MOB-1704: extract the header's subject label, thumbnail, and location line
Moves subjectLabel, SubjectThumbnail, and LocationSubtitle out of
ExploreV2Header verbatim so saved search rows can show the same thing
the header does. No behavior change; testIDs are unchanged.
